- validation added

+++ app/src/main/java/com/example/parth_c0766346_fp/Manager.java
@@ -7,7 +7,7 @@ public class Manager extends Employee {
     private final int GAIN_FACTOR_CLIENT = 500;
     private int nbClients;
 
-    public Manager(String name, int id, int birthYear, double monthlySalary, double rate, Vehicle empVehicle, int nbClients) {
+    public Manager(String name, int id, int birthYear, double monthlySalary, int rate, Vehicle empVehicle, int nbClients) {
         super(name, id, birthYear, monthlySalary, rate, empVehicle);
         this.nbClients = nbClients;
     }
