+++ app/src/main/java/com/example/parth_c0766346_fp/Programmer.java
@@ -7,7 +7,7 @@ public class Programmer extends Employee {
     private final int GAIN_FACTOR_PROJECTS = 200;
     private int nbProjects;
 
-    public Programmer(String name, int id, int birthYear, double monthlySalary, double rate, Vehicle empVehicle, int nbProjects) {
+    public Programmer(String name, int id, int birthYear, double monthlySalary, int rate, Vehicle empVehicle, int nbProjects) {
         super(name, id, birthYear, monthlySalary, rate, empVehicle);
         this.nbProjects = nbProjects;
     }
