+++ app/src/main/java/com/example/parth_c0766346_fp/Tester.java
@@ -6,7 +6,7 @@ public class Tester extends Employee {
     private final int GAIN_FACTOR_ERROR = 10;
     int nbBugs;
 
-    public Tester(String name, int id, int birthYear, double monthlySalary, double rate, Vehicle empVehicle, int nbBugs) {
+    public Tester(String name, int id, int birthYear, double monthlySalary, int rate, Vehicle empVehicle, int nbBugs) {
         super(name, id, birthYear, monthlySalary, rate, empVehicle);
         this.nbBugs = nbBugs;
     }
